+++ src/policy/handcrafted/spell_profile.rs
@@ -513,6 +513,7 @@ impl HandcraftedPolicy {
             | ValueDef::CountersOnSource(_)
             | ValueDef::PaidAmount
             | ValueDef::MatchedCount
+            | ValueDef::BoundObjectCount(_)
             | ValueDef::SpellsCastBeforeThisTurn
             | ValueDef::DividedAmongTargets
             | ValueDef::TargetPower(_)